Progress on analyses
checked basic statistics, added those details to the code
Started on analyses

Next steps:
Need to create a fuller set of output files for different analyses - the intra and interspecific niche analyses are done on the dataset with all isotopes, even samples with really old dates, so need to deal with this.
Once I do this, I should move the basic stats to the Analysis code scripts.